当 mount.ntfs3存在时，不采用ntfs-3g

+++ mk_rk3328_beikeyun.sh
@@ -443,7 +443,7 @@ ln -sf kmod lsmod
 ln -sf kmod modinfo
 ln -sf kmod modprobe
 ln -sf kmod rmmod
-ln -sf /usr/bin/ntfs-3g mount.ntfs
+[ ! -f mount.ntfs3 ] && [ -f ../usr/bin/ntfs-3g ] && ln -sf /usr/bin/ntfs-3g mount.ntfs
 
 cd $TGT_ROOT/lib/firmware
 mv *.hcd brcm/ 2>/dev/null
